"encoding/gob" is about 2x faster than "encoding/binary".
But adding GobEncode() and GobDecode() interfaces gives same performance as
"encoding/binary". So `bkey` embedded structure in `block` had to be removed
and replaced with slice of keys, slice of docids, slice of values. This lead
to replacing all kn.ks[index].dpos and kn.ks[index].kpos with kn.ks[index] and
kn.ds[index] inside the algorithm. Benchmark cases are added in `block_test`,
try,

$ go test -test.bench="gob"

So far `ctrl` field in `bkey` did not prove useful, so removing that field.